Move virtualfiles to cocoasm module.
