+++ .travis.yml
@@ -4,6 +4,6 @@ language: python
 python:
   - 3.6
 script:
-  - nosetests -v --with-coverage --cover-package=cocoasm --cover-package=fileutil
+  - nosetests -v --with-coverage --cover-package=cocoasm
 after_success:
   - bash <(curl -s https://codecov.io/bash) || echo "Codecov did not collect coverage reports"
