+++ cocoasm/virtualfiles.py
